closes #3. Rearange history and song info panels.
