+++ README.md
@@ -1,2 +1,3 @@
-# Radio Substream
+# :metro: Radio Sub Stream
 Application that joins radio streams from multiple servers.
+
